Crate zng_wgt_material_icons

Material icons for the Icon! widget.

A map from name to icon codepoint is defined in a module for each font. The font files are embedded by default and can are registered using the MaterialIconsManager app extension. The extension also registers ICONS handlers that provide the icons.

The icons are from the Material Design Icons project.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 5 feature flags, 5 enabled by default.

"embedded"

Embedded font files.

Enabled by default.

"outlined"

Outlined icon set.

Enabled by default.

"filled"

Filled icon set.

Enabled by default.

"rounded"

Rounded icon set.

Enabled by default.

"sharp"

Sharp icon set.

Enabled by default.

Modules

filled

Filled icons.

outlined

Outline icons.

rounded

Rounded icons.

sharp

Sharp icons.

Structs

MaterialIconsManager

Material icon fonts manager.